+++ Final-Project-GameStore/src/test/java/com/company/gamestore/repository/FeeRepositoryTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -20,7 +21,7 @@ public class FeeRepositoryTest {
     @Test
     public void shouldGetFeeByProductType(){
         Fee fee = new Fee();
-        fee.setFee(14.99);
+        fee.setFee(BigDecimal.valueOf(14.99));
         fee.setProductType("Console");
 
         List<Fee> fees = feeRepository.findByProductType("Console");
