+++ Final-Project-GameStore/src/test/java/com/company/gamestore/repository/InvoiceRepositoryTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -30,12 +31,12 @@ public class InvoiceRepositoryTest {
         invoice.setZipcode("94065");
         invoice.setItemType("Console");
         invoice.setItemId(1412);
-        invoice.setUnitPrice(49.99);
+        invoice.setUnitPrice(BigDecimal.valueOf(49.99));
         invoice.setQuantity(1);
-        invoice.setSubtotal(49.99);
-        invoice.setTax(2.99);
-        invoice.setProcessingFee(1.99);
-        invoice.setTotal(54.97);
+        invoice.setSubtotal(BigDecimal.valueOf(49.99));
+        invoice.setTax(BigDecimal.valueOf(2.99));
+        invoice.setProcessingFee(BigDecimal.valueOf(1.99));
+        invoice.setTotal(BigDecimal.valueOf(54.97));
         invoiceRepository.save(invoice);
 
     }
@@ -63,12 +64,12 @@ public class InvoiceRepositoryTest {
         invoice1.setZipcode("95032");
         invoice1.setItemType("Console");
         invoice1.setItemId(1412);
-        invoice1.setUnitPrice(49.99);
+        invoice1.setUnitPrice(BigDecimal.valueOf(49.99));
         invoice1.setQuantity(2);
-        invoice1.setSubtotal(99.98);
-        invoice1.setTax(5.99);
-        invoice1.setProcessingFee(1.99);
-        invoice1.setTotal(107.96);
+        invoice1.setSubtotal(BigDecimal.valueOf(99.98));
+        invoice1.setTax(BigDecimal.valueOf(5.99));
+        invoice1.setProcessingFee(BigDecimal.valueOf(1.99));
+        invoice1.setTotal(BigDecimal.valueOf(107.96));
         invoiceRepository.save(invoice1);
         List<Invoice> invoices = invoiceRepository.findAll();
         assertTrue(invoices.contains(invoice1));
@@ -78,13 +79,13 @@ public class InvoiceRepositoryTest {
 
     @Test
     public void shouldUpdateInvoice(){
-        invoice.setProcessingFee(2.99);
-        invoice.setTotal(55.97);
+        invoice.setProcessingFee(BigDecimal.valueOf(2.99));
+        invoice.setTotal(BigDecimal.valueOf(55.97));
         invoiceRepository.save(invoice);
         Optional<Invoice> invoice1 = invoiceRepository.findById(invoice.getInvoiceId());
         assertEquals(invoice1.get(), invoice);
-        assertEquals(invoice1.get().getProcessingFee(), 2.99);
-        assertEquals(invoice1.get().getTotal(), 55.97);
+        assertEquals(invoice1.get().getProcessingFee(), BigDecimal.valueOf(2.99));
+        assertEquals(invoice1.get().getTotal(), BigDecimal.valueOf(55.97));
     }
 
     @Test
